Refactor Supabase client modules
