+++ app/admin/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from 'next/navigation';
 import { getProfile } from '@/lib/auth';
-import { createSupabaseServerClient } from '@/lib/supabase';
+import { createSupabaseServerClient } from '@/lib/supabase-server';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card';
 import { revalidatePath } from 'next/cache';
