+++ lib/supabase-browser.ts
@@ -1,0 +1,9 @@
+import { createBrowserClient } from '@supabase/auth-helpers-nextjs';
+import { type Database } from '@/types/database';
+
+export function createSupabaseBrowserClient() {
+  return createBrowserClient<Database>(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
+  );
+}
