+++ lib/supabase.ts
@@ -1,32 +1,2 @@
-import { cookies } from 'next/headers';
-import { createBrowserClient, createServerClient } from '@supabase/auth-helpers-nextjs';
-import { type CookieOptions } from '@supabase/auth-helpers-shared';
-import { type Database } from '@/types/database';
-
-export function createSupabaseServerClient() {
-  const cookieStore = cookies();
-  return createServerClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-      cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
-        },
-        set(name: string, value: string, options: CookieOptions) {
-          cookieStore.set({ name, value, ...options });
-        },
-        remove(name: string, options: CookieOptions) {
-          cookieStore.set({ name, value: '', ...options });
-        }
-      }
-    }
-  );
-}
-
-export function createSupabaseBrowserClient() {
-  return createBrowserClient<Database>(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-  );
-}
+export { createSupabaseServerClient } from './supabase-server';
+export { createSupabaseBrowserClient } from './supabase-browser';
